Added program

+++ com/heraizen/cj/day1/SumOfFactorialSeries.java
@@ -3,9 +3,9 @@ package com.heraizen.cj.day1;
 public class SumOfFactorialSeries {
 
 	public static void main(String[] args) {
-		int sum = 0;
+		float sum = 0;
 		for (int i = 1; i <= 7; i++) {
-			sum += i / factorial(i);
+			sum += (float) i / factorial(i);
 		}
 		System.out.println("sum of series : " + sum);
 	}
